Revert "Fix priority-based downlink policy default behavior (#1470)"
This reverts commit 6feb3c1f1ee0a436d95f9cd0eaf8ba502448dc8e.

In the case of simulcast, this fix seems to *cause* the exact behaviour
it perports to fix.

We're using Simulcast, and relying on the default behaviour to prioritise all streams equally.
This was working fine before. It broke now.

Reverting this patch fixes the behaviour for us.

After debugging this, the root of the issue seems to be the logic where "videoPreferences"
is set to "undefined."

Then, the first time VideoPriorityBasedPolicy:updateIndex is run, it sets video preferences
to 'defaultVideoPreferences' - After which, videoPreferences is fixed to the initial set
of streams in the first index. It will never again re-calculate a default set as
new participants join, as it will always check for the existence of videoPreferences,
and since it exists, will not update it.
This is different behaviour to what was happening prior to this patch, where the default
preference set was recalculated with each index update.